% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{rdf_serialize}
\alias{rdf_serialize}
\title{Serialize RDF Data}
\usage{
rdf_serialize(store, format = "turtle")
}
\arguments{
\item{store}{An RDF store handle}

\item{format}{RDF format: "turtle", "ntriples", "rdfxml", "nquads", or "trig"}
}
\value{
The serialized RDF data as a character string
}
\description{
Serializes the store contents to a string.
}
\examples{
store <- rdf_store()
rdf_load(store, '<http://example.org/s> <http://example.org/p> "value" .', format = "ntriples")
rdf_serialize(store, format = "turtle")
}
