% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{rdf_size}
\alias{rdf_size}
\title{Get Store Size}
\usage{
rdf_size(store)
}
\arguments{
\item{store}{An RDF store handle}
}
\value{
The number of quads as an integer
}
\description{
Returns the number of quads (triples) in the store.
}
\examples{
store <- rdf_store()
rdf_size(store)
}
