% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdf_store_insert}
\alias{rdf_store_insert}
\title{Insert a triple into the store}
\usage{
rdf_store_insert(store_idx, subject, predicate, object, graph)
}
\arguments{
\item{store_idx}{Store index}

\item{subject}{Subject IRI (e.g., "\url{http://example.org/s}") or blank node ("_:b1")}

\item{predicate}{Predicate IRI (e.g., "\url{http://example.org/p}")}

\item{object}{Object (IRI, blank node, or literal with quotes e.g., "\"value\"")}

\item{graph}{Optional graph name IRI}
}
\description{
Insert a triple into the store
}
