% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdf_store_load}
\alias{rdf_store_load}
\title{Load RDF data into the store}
\usage{
rdf_store_load(store_idx, data, format, base_iri)
}
\arguments{
\item{store_idx}{Store index}

\item{data}{RDF data as a string}

\item{format}{RDF format: "turtle", "ntriples", "rdfxml", "nquads", "trig"}

\item{base_iri}{Optional base IRI for relative URIs}
}
\description{
Load RDF data into the store
}
