% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdf_store_new}
\alias{rdf_store_new}
\title{Create a new in-memory RDF store}
\usage{
rdf_store_new()
}
\value{
Store index (integer handle)
}
\description{
Create a new in-memory RDF store
}
\section{Arguments}{

This function takes no arguments.
}

\examples{
store <- rdf_store_new()
rdf_store_size(store)
}
\keyword{internal}
