% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdf_store_open}
\alias{rdf_store_open}
\title{Open or create a persistent RDF store at the given path}
\usage{
rdf_store_open(path)
}
\arguments{
\item{path}{Path to the store directory}
}
\value{
Store index (integer handle)
}
\description{
Open or create a persistent RDF store at the given path
}
\examples{
\donttest{
if (.Platform$OS.type != "windows") {
  store <- rdf_store_open(file.path(tempdir(), "roxigraph_test"))
}
}
}
\keyword{internal}
