% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdf_store_query}
\alias{rdf_store_query}
\title{Execute a SPARQL query and return results as a data frame}
\usage{
rdf_store_query(store_idx, query)
}
\arguments{
\item{store_idx}{Store index}

\item{query}{SPARQL query string}
}
\value{
Query results as a data frame (for SELECT) or logical (for ASK)
}
\description{
Execute a SPARQL query and return results as a data frame
}
\examples{
store <- rdf_store_new()
rdf_store_load(store, '<http://example.org/s> <http://example.org/p> "v" .', "ntriples", NULL)
rdf_store_query(store, "SELECT * WHERE { ?s ?p ?o }")
}
\keyword{internal}
