% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdf_store_remove}
\alias{rdf_store_remove}
\title{Remove a triple from the store}
\usage{
rdf_store_remove(store_idx, subject, predicate, object, graph)
}
\arguments{
\item{store_idx}{Store index}

\item{subject}{Subject IRI or blank node}

\item{predicate}{Predicate IRI}

\item{object}{Object}

\item{graph}{Optional graph name IRI}
}
\description{
Remove a triple from the store
}
