% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdf_store_size}
\alias{rdf_store_size}
\title{Get the number of quads in the store}
\usage{
rdf_store_size(store_idx)
}
\arguments{
\item{store_idx}{Store index}
}
\value{
The number of quads
}
\description{
Get the number of quads in the store
}
\examples{
store <- rdf_store_new()
rdf_store_size(store)
}
\keyword{internal}
