% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{sparql_update}
\alias{sparql_update}
\title{Execute SPARQL Update}
\usage{
sparql_update(store, update)
}
\arguments{
\item{store}{An RDF store handle}

\item{update}{A SPARQL UPDATE query string}
}
\value{
Invisibly returns NULL
}
\description{
Executes a SPARQL UPDATE query to modify the store.
}
\examples{
store <- rdf_store()
sparql_update(store, "INSERT DATA { <http://example.org/s> <http://example.org/p> 'value' }")
}
