% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdf_store_dump}
\alias{rdf_store_dump}
\title{Serialize the store contents to a string}
\usage{
rdf_store_dump(store_idx, format)
}
\arguments{
\item{store_idx}{Store index}

\item{format}{RDF format: "turtle", "ntriples", "rdfxml", "nquads", "trig"}
}
\value{
The serialized RDF data
}
\description{
Serialize the store contents to a string
}
\examples{
store <- rdf_store_new()
rdf_store_load(store, '<http://example.org/s> <http://example.org/p> "v" .', "ntriples", NULL)
rdf_store_dump(store, "turtle")
}
\keyword{internal}
