% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdf_store_remove}
\alias{rdf_store_remove}
\title{Remove a triple from the store}
\usage{
rdf_store_remove(store_idx, subject, predicate, object, graph)
}
\arguments{
\item{store_idx}{Store index}

\item{subject}{Subject IRI or blank node}

\item{predicate}{Predicate IRI}

\item{object}{Object}

\item{graph}{Optional graph name IRI}
}
\value{
No return value, called for side effects (removing triples from the store)
}
\description{
Remove a triple from the store
}
\examples{
store <- rdf_store_new()
rdf_store_insert(store, "<http://example.org/s>", "<http://example.org/p>", '"val"', NULL)
rdf_store_remove(store, "<http://example.org/s>", "<http://example.org/p>", '"val"', NULL)
}
\keyword{internal}
