% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rdf_store_update}
\alias{rdf_store_update}
\title{Execute a SPARQL UPDATE query}
\usage{
rdf_store_update(store_idx, update)
}
\arguments{
\item{store_idx}{Store index}

\item{update}{SPARQL UPDATE query string}
}
\value{
No return value, called for side effects (executing SPARQL UPDATE queries)
}
\description{
Execute a SPARQL UPDATE query
}
\examples{
store <- rdf_store_new()
rdf_store_update(store, "INSERT DATA { <http://example.org/s> <http://example.org/p> 'val' }")
}
\keyword{internal}
